% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogimet.R
\name{direct_download_from_ogimet}
\alias{direct_download_from_ogimet}
\title{Direct download of meteorological data from Ogimet}
\usage{
direct_download_from_ogimet(wmo_identifier, initial_date, final_date)
}
\arguments{
\item{wmo_identifier}{A 5-digit character string or integer representing the station WMO ID.}

\item{initial_date}{Initial date, format "YYYY-MM-DD".}

\item{final_date}{Final date, format "YYYY-MM-DD".}
}
\value{
A data frame, as returned by \code{show_synop_data()}
}
\description{
Direct download of meteorological data from Ogimet
}
\details{
The requested period cannot exceed 370 days. All queries assume UTC time zone.
The returned data frame covers from 00:00 UTC of the \code{initial_date} to 23:00 UTC
of the \code{final_date}, inclusive.
Too many requests may trigger temporal blocks.

If the station identifier starts with 0 (zero), then \code{wmo_identifier} must be a string (e.g., "06447").
}
\examples{
\dontrun{
direct_download_from_ogimet(wmo_identifier = '87585',
                            initial_date = "2024-01-10",
                            final_date = "2024-01-11")
}

}
