% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogimet.R
\name{download_from_ogimet}
\alias{download_from_ogimet}
\title{Download SYNOP messages from Ogimet}
\usage{
download_from_ogimet(wmo_identifier, initial_date, final_date)
}
\arguments{
\item{wmo_identifier}{A 5-digit character string or integer representing the station WMO ID.}

\item{initial_date}{Initial date, format "YYYY-MM-DD".}

\item{final_date}{Final date, format "YYYY-MM-DD".}
}
\value{
A character vector with SYNOP strings.
}
\description{
Download SYNOP messages from Ogimet
}
\details{
The requested period cannot exceed 370 days. All queries assume UTC time zone.
The returned dataset covers from 00:00 UTC of the \code{initial_date} to 23:00 UTC
of the \code{final_date}, inclusive.
}
\examples{
\donttest{
download_from_ogimet(wmo_identifier = 87585, initial_date = "2024-01-10", final_date = "2024-01-11")
}

}
