% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\name{parse_ogimet}
\alias{parse_ogimet}
\title{Parse Ogimet CSV strings into a data frame}
\usage{
parse_ogimet(ogimet_data)
}
\arguments{
\item{ogimet_data}{A character vector of Ogimet strings.}
}
\value{
A tibble with Year, Month, Day, Hour, and Raw_Synop.
}
\description{
Parse Ogimet CSV strings into a data frame
}
