% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sy_nearest_station.R
\name{sy_nearest_station}
\alias{sy_nearest_station}
\title{sy_nearest_station}
\usage{
sy_nearest_station(x)
}
\arguments{
\item{x}{object of class sfg, sfc or sf}
}
\value{
sf object
}
\description{
sy_nearest_station
}
\examples{
\dontrun{
paris <- st_sfc(st_point(c(2.335, 48.876)), crs = 4326)
nearest_station <- sy_nearest_station(paris)
}

}
