% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sy_plot.R
\name{sy_plot_temp}
\alias{sy_plot_temp}
\title{sy_plot_temp}
\usage{
sy_plot_temp(synop, station)
}
\arguments{
\item{synop}{data.frame; result of \code{sy_download()}}

\item{station}{numeric; id of station}
}
\value{
plot
}
\description{
sy_plot_temp
}
\examples{
\dontrun{
# download data
synop <- sy_download(date = c("202201", "202304"), T)

# find nearest station
paris <- st_sfc(st_point(c(2.335, 48.876)), crs = 4326)
nearest_station <- sy_nearest_station(paris)
num_station <- as.numeric(nearest_station$ID)

# plot temperature
sy_plot_temp(synop, num_station)
}

}
