% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogimet.R
\name{parse_ogimet}
\alias{parse_ogimet}
\title{Parse Ogimet strings into a data frame}
\usage{
parse_ogimet(ogimet_data)
}
\arguments{
\item{ogimet_data}{A character vector of Ogimet strings.}
}
\value{
A tibble with Year, Month, Day, Hour, and Raw_synop.
}
\description{
Parse Ogimet strings into a data frame
}
\examples{
msg <- paste0("87736,2026,01,01,12,00,AAXX 01123 87736 32965 13205 10214 20143 ",
              "30022 40113 5//// 80005 333 10236 20128=")
parsed_data <- parse_ogimet(msg)
}
