% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sy_download.R
\name{sy_download}
\alias{sy_download}
\title{sy_download}
\usage{
sy_download(date, between = F)
}
\arguments{
\item{date}{character; vector of dates format as "YYYYMM"}

\item{between}{boolean; if TRUE all month between first and last date are downloaded.
In this case, \code{date} must be set with only two date.}
}
\value{
data.frame
}
\description{
sy_download
}
\examples{
\dontrun{
dates <- c("202201", "202301")
two_month <- sy_download(dates)
all_year <- sy_download(dates, between = T)

}

}
